:sparkles: ((backend)): Add `since` and `until` params to /views endpoint
Implemented a dummy date range filter after we query all the data, to be change in the future for a more optimal filter inside the query. This is just enough to allow me to work forward with the frontend. Also add since and until params with default value in the frontend query.