:sparkles: (frontend): add dummy date range picker from rsuite
Until cunningam doesn't provide a date range picker component we will use the one from the rsuite package. It is define in packages/ui so we can change it later without changing the code in apps/web. This commit just add the date picker but is doesn't change the query date range.