✨(api) add date/time range filter
As we manipulate time-based data, we need a way to filter queries using
date/time ranges. Current filter is generic and should be injected as a
dependency for every route corresponding to an indicator. A first
example integration has been made for the daily video views.

One should note that:

1. default time range corresponds to the last seven days,
2. the time range cannot be greater than the API_MAX_TIMESPAN setting.