👷(backend) use astral-sh/ruff to make linting faster and simpler
Introduce ruff and use it to replace isort, flake8, pycodestyle and
bandit by configuring it through pyproject.toml

This change might not be iso but should be close and brings a
considerable speed improvement to the make lint command

Not all checks have been activated in the config file so we can still
make ruff stricter.

The fix flag is set to true and ruff will (apparently aggressively) fix
the violations. This behavior can be tamed via the unfixable config
entry