🔧(frontend) render the Web app with the Django LTI server
Replaced the conventional development server with a process that builds the
frontend project, and then stores its output within Django's static files.

Django injects the JS bundle and the stylesheet in its base template. SPA
wait for the DOM to be mounted, and thus render the app in a dedicated div.
This allows Django server to write data to the DOM, thus making accessible
data to the SPA, that parses the DOM.

Additionally, the allowed host has been updated to reflect the Django server
port. This adjustment is vital for seamless API calls from the frontend in the
future.