🩹(api) fix client error handling with HttpException
Integrated the default HttpException to enhance our API's error handling
capabilities. The HttpException behaves like a standard Python exception
with additional data relevant for APIs.

When this exception is raised, it effectively prevents the further execution
of the API call. We unlock the potential for more advanced configuration
options during exception handling.

Moreover, this approach eliminates the need for custom Pydantic models
dedicated to managing API responses, leading to a  more maintainable code.