👷(ci) harmonize CI linting rules with local development guidelines
The `lint-git` CI job might not be the optimal location for examining print
statements, as this is more aligned with the responsibility of our linter.
Moreover, the inspection of print statements is language-specific and
should pertain only to jobs running on Python files.

`lint-git` is running both on python and JavaScript files.