✨(api) add indicator persistence mixins
We need a generic way to cache indicators calculation so that we can
speed up API responses. Caching indicator results should be performed
asynchronously or pre-calculated periodically (as cron jobs).

Our proposition relies on indicators mixins that can be added to
standard indicators to add cache support. We distinguish two types of
cache: a dummy cache for metrics that mostly rely on a single result
that cannot be aggregated periodically, and an incremental cache for
indicators that can be computed on a defined time span and aggregated on
a larger time span.