🐛(frontend) prevent overwriting app state with the dataContext
When `appData` updates occurred, the `AppContentLoader` component's
rendering was triggered. This rendering action was causing a reset
of the `appData` to its initial state, resulting in the unintended
overwriting of crucial changes, including the recently refreshed
access token.

Now, the AppContentLoader initializes the data provided by Django only during
the initial rendering to avoid unintended overwrites.