🚑️(frontend) configure API and APP root urls in production
Previously, the APIs root URLs were statically resolved during the frontend
build using Vite. This meant that changing these URLs required a rebuild of
the frontend image for a new user.

The frontend environment variables are designed to be overridden by values
passed through the Django template. This design choice enables the specification
of APIs root URLs in production without necessitating a rebuild of the
frontend image.