🐛(project) define default doc version as development
As we use Mike to support multiple documentation versions and we only
have a single documentation corresponding to the main branch, we need to
set the default documentation target so that it points to the main
branch and is labelled as `dev`.

We also improved the development environment to check multiple
documentation versions build using Mike locally.