📝(project) improve mike local usage
Our CI bot should not be the only user allowed to act on the
documentation. If we need to perform clean up actions from our
workstation, Mike needs to be able to commit on Warren's repository on
our behalf.

Hence the documentation Docker image now runs using a dedicated user
that shares the Git and SSH configurations from the host.