🚨(helm) fix linter warning for empty `allowedHosts` value
On liveness and readiness probes, with virtual hosts, `Host` should be set in
the Host header. As the default value is an empty list, Helm linter was raising
an error.
Fixing the lint issue by only adding an `httpHeaders` if value `allowHosts` is
not empty, for both `api` and `app` deployments.