✨(api) change cache key to indicator attributes
Previously, variations in indicator attributes such as boolean flags or
environment settings led to divergent results for identical LRS queries.
We could not use cache mixins in the cases, as we only rely on the LRS query
as a cache key.
Changing the cache key to be a hash from the indicator instance attributes, so
that cache mixins are usable for these indicators.