🔨(api) make migrations with warren cli
In PR #249, the default database URL was removed from the Alembic configuration
file to facilitate testing with the test database without performing any
migrations on the real database.
However, this change prevented migrations from being executed when using the
Alembic binary, causing the bootstrap makefile rule to fail.
Adding the possibility to execute migrations when warren-api is not running.