🔨(api) fix warren development cli script
Warren CLI offers commands for migrating the database (we want these to be
runnable even when the API is not running), but also offers commands for
indexing courses and content of an LMS, or for computing indicators (we want
these to be executed alongside a running API instance).
Adding the possibility to run those two types on commands in our docker script.