🔒️(project) securely pass user and course ids through jwt
Previously, `course_id` was transmitted to the frontend through app context
data.
`user_id` was not transmitted to the frontend.
To be secure, both `course_id` and `user_id` will now be transmitted through the
LTI token instead of the app context data.
This change prevents users from altering the course_id or user_id to access
unauthorized data.