✨(xi) make timeout of requests to Moodle configurable
The default timeout for requests to retrieve courses and content from Moodle was
set to 5 seconds (httpx's default).
This duration can be too short for Moodle instances with a large number of
courses.
Allowing the timeout to be configurable, with a new default value of 60 seconds,
which should be plenty enough for most LMS instances.